# docs/swagger.yaml
basePath: /
definitions:
...
  camera.CapturedEventData:
    properties:
      ChannelName:
        type: string
      EventComment:
        type: string
      EventDescription:
        type: string
      EventId:
        type: string
      captured_time:
        type: string
    type: object
  modelscar.Car_Model:
    properties:
      car_number:
        type: string
      duration:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      image_url:
        type: string
      park_no:
        type: string
      reason:
        type: string
      start_time:
        type: string
      status:
        type: string
      total_payment:
        type: number
      user_id:
        type: string
    type: object
...
  modelsuser.User:
    type: object
...
  resmodel.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  resmodel.Response:
    properties:
      data:
        $ref: '#/definitions/modelscar.Car_Model'
      message:
        type: string
    type: object
  usercontrol.LoginInput:
    properties:
      parkno:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
host: 192.168.100.7:3000
info:
  contact: {}
  title: Airline REST API
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: '{ "username": "Dowran", "password": "12345678", "parkno": "P4"
        }'
      parameters:
      - description: User Login Data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/usercontrol.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Login successful'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'message: Invalid request body'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Invalid username or password'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login User
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Ends the session of a logged-in user by deleting the JWT token
        cookie.
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Logout successful'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout User
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Retrieves the current user's username, role, and user ID from the
        JWT token.
      produces:
      - application/json
      responses:
        "200":
          description: Returns user information
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'message: Bad Request - Missing data from middleware'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'message: Unauthorized - Invalid token'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal Server Error - Missing data from middleware'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user information
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 'Creates a new user and stores their hashed password. Example:
        { "username": "newUser", "password": "password123", "firstname": "John", "lastname":
        "Doe", "role": "admin" }'
      parameters:
      - description: User Registration Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/modelsuser.User'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: User Created'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'message: Password must be at least 8 characters long'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register User
      tags:
      - Auth
  /api/v1/camera/getdata:
    post:
      consumes:
      - application/json
      description: '{"ChannelName": "P41", "EventComment": "BE5084AG"}'
      parameters:
      - description: Captured data from the camera
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/camera.CapturedEventData'
      produces:
      - application/json
      responses:
        "201":
          description: Car entry created successfully
          schema:
            $ref: '#/definitions/resmodel.Response'
        "400":
          description: Bad request, car is already inside
          schema:
            $ref: '#/definitions/resmodel.ErrorResponse'
        "500":
          description: Internal server error, failed to save data
          schema:
            $ref: '#/definitions/resmodel.ErrorResponse'
      summary: Create a new car entry in the parking lot
      tags:
      - Car Entry
    put:
      consumes:
      - application/json
      description: '{"EventComment": "BE5084AG"}'
      parameters:
      - description: Captured data from the camera
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/camera.CapturedEventData'
      produces:
      - application/json
      responses:
        "200":
          description: Car exit updated successfully
          schema:
            $ref: '#/definitions/resmodel.Response'
        "400":
          description: Bad request, car already exited
          schema:
            $ref: '#/definitions/resmodel.ErrorResponse'
        "404":
          description: Car not found
          schema:
            $ref: '#/definitions/resmodel.ErrorResponse'
        "500":
          description: Internal server error, failed to update data
          schema:
            $ref: '#/definitions/resmodel.ErrorResponse'
      summary: Create a car exit record in the parking lot
      tags:
      - Car Entry
...
swagger: "2.0"